add file sort and automatic directory expansion
