+++ src/components/File/FileSideBar/FileItem.tsx
@@ -1,0 +1,113 @@
+import { removeFileExtension } from "@/functions/strings";
+import { FileInfoNode } from "electron/main/Files/Types";
+import React, { useState } from "react";
+import { FaChevronRight } from "react-icons/fa";
+import { FaChevronDown } from "react-icons/fa";
+import { isFileNodeDirectory, moveFile } from "./fileOperations";
+
+interface FileInfoProps {
+  file: FileInfoNode;
+  selectedFilePath: string | null;
+  onFileSelect: (path: string) => void;
+  handleDragStart: (e: React.DragEvent, file: FileInfoNode) => void;
+  onDirectoryToggle: (path: string) => void;
+  isExpanded?: boolean;
+  indentMultiplyer?: number;
+}
+
+export const FileItem: React.FC<FileInfoProps> = ({
+  file,
+  selectedFilePath,
+  onFileSelect,
+  handleDragStart,
+  onDirectoryToggle,
+  isExpanded,
+  indentMultiplyer,
+}) => {
+  const isDirectory = isFileNodeDirectory(file);
+  const isSelected = file.path === selectedFilePath;
+  const indentation = indentMultiplyer ? 10 * indentMultiplyer : 0;
+  const [isDragOver, setIsDragOver] = useState(false);
+
+  const handleDragOver = (e: React.DragEvent) => {
+    e.preventDefault();
+    setIsDragOver(true);
+  };
+
+  const handleDragLeave = () => {
+    setIsDragOver(false);
+  };
+
+  const handleDrop = async (e: React.DragEvent) => {
+    e.preventDefault();
+    setIsDragOver(false); // Reset drag over state
+    const sourcePath = e.dataTransfer.getData("text/plain");
+    let destinationPath = file.path; // Default destination path is the path of the file item itself
+
+    if (!isFileNodeDirectory(file)) {
+      const pathSegments = file.path.split("/");
+      pathSegments.pop(); // Remove the file name from the path
+      destinationPath = pathSegments.join("/");
+    }
+
+    try {
+      moveFile(sourcePath, destinationPath);
+      // Refresh file list here or in moveFile function
+    } catch (error) {
+      console.error("Failed to move file:", error);
+      // Handle error (e.g., show an error message)
+    }
+  };
+  const toggle = () => {
+    if (isFileNodeDirectory(file)) {
+      onDirectoryToggle(file.path);
+    } else {
+      onFileSelect(file.path);
+    }
+  };
+
+  const localHandleDragStart = (e: React.DragEvent) => {
+    e.stopPropagation();
+    handleDragStart(e, file);
+  };
+
+  const handleContextMenu = (e: React.MouseEvent) => {
+    e.preventDefault();
+    window.contextMenu.showFileItemContextMenu(file);
+  };
+
+  const itemClasses = `flex items-center cursor-pointer px-2 py-1 border-b border-gray-200 hover:bg-slate-700 h-full mt-0 mb-0 ${
+    isSelected ? "bg-gray-700 text-white font-semibold" : "text-gray-200"
+  } ${isDragOver ? "bg-blue-500" : ""}`;
+
+  return (
+    <div
+      draggable
+      onDragStart={localHandleDragStart}
+      onContextMenu={handleContextMenu}
+      style={{ paddingLeft: `${indentation}px` }}
+      onDragOver={handleDragOver}
+      onDrop={handleDrop}
+      onDragLeave={handleDragLeave}
+    >
+      <div onClick={toggle} className={itemClasses}>
+        {isDirectory && (
+          <span className={`mr-2 mt-1 text-[13px] `}>
+            {isExpanded ? (
+              <FaChevronDown title="Collapse Directory" />
+            ) : (
+              <FaChevronRight title="Open Directory" />
+            )}
+          </span>
+        )}
+        <span
+          className={`text-[13px] flex-1 truncate mt-0 ${
+            isDirectory ? "font-semibold" : ""
+          }`}
+        >
+          {removeFileExtension(file.name)}
+        </span>
+      </div>
+    </div>
+  );
+};
